+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_global_endpoint_manager_async.py
@@ -24,8 +24,12 @@ database service.
 """
 
 import asyncio # pylint: disable=do-not-import-asyncio
+import logging
+from asyncio import CancelledError
+from typing import Tuple
 
 from azure.core.exceptions import AzureError
+from azure.cosmos import DatabaseAccount
 
 from .. import _constants as constants
 from .. import exceptions
@@ -34,7 +38,9 @@ from .._location_cache import LocationCache
 
 # pylint: disable=protected-access
 
-class _GlobalEndpointManager(object):
+logger = logging.getLogger("azure.cosmos._GlobalEndpointManager")
+
+class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attributes
     """
     This internal class implements the logic for endpoint management for
     geo-replicated database accounts.
@@ -53,6 +59,8 @@ class _GlobalEndpointManager(object):
             client.connection_policy.UseMultipleWriteLocations,
             self.refresh_time_interval_in_ms,
         )
+        self.startup = True
+        self.refresh_task = None
         self.refresh_needed = False
         self.refresh_lock = asyncio.Lock()
         self.last_refresh_time = 0
@@ -62,13 +70,10 @@ class _GlobalEndpointManager(object):
         return constants._Constants.DefaultUnavailableLocationExpirationTime
 
     def get_write_endpoint(self):
-        return self.location_cache.get_write_regional_endpoint()
+        return self.location_cache.get_write_regional_routing_context()
 
     def get_read_endpoint(self):
-        return self.location_cache.get_read_regional_endpoint()
-
-    def swap_regional_endpoint_values(self, request):
-        return self.location_cache.swap_regional_endpoint_values(request)
+        return self.location_cache.get_read_regional_routing_context()
 
     def resolve_service_endpoint(self, request):
         return self.location_cache.resolve_service_endpoint(request)
@@ -85,14 +90,21 @@ class _GlobalEndpointManager(object):
     def can_use_multiple_write_locations(self, request):
         return self.location_cache.can_use_multiple_write_locations_for_request(request)
 
-    async def force_refresh(self, database_account):
+    async def force_refresh_on_startup(self, database_account):
         self.refresh_needed = True
         await self.refresh_endpoint_list(database_account)
+        self.startup = False
 
     def update_location_cache(self):
         self.location_cache.update_location_cache()
 
     async def refresh_endpoint_list(self, database_account, **kwargs):
+        if self.refresh_task and self.refresh_task.done():
+            try:
+                await self.refresh_task
+                self.refresh_task = None
+            except (Exception, CancelledError) as exception: #pylint: disable=broad-exception-caught
+                logger.exception("Health check task failed: %s", exception)
         if self.location_cache.current_time_millis() - self.last_refresh_time > self.refresh_time_interval_in_ms:
             self.refresh_needed = True
         if self.refresh_needed:
@@ -114,46 +126,75 @@ class _GlobalEndpointManager(object):
             if self.location_cache.should_refresh_endpoints() or self.refresh_needed:
                 self.refresh_needed = False
                 self.last_refresh_time = self.location_cache.current_time_millis()
-                database_account = await self._GetDatabaseAccount(**kwargs)
-                self.location_cache.perform_on_database_account_read(database_account)
-                # this will perform getDatabaseAccount calls to check endpoint health
-                await self._endpoints_health_check(**kwargs)
+                if not self.startup:
+                    # this will perform getDatabaseAccount calls to check endpoint health
+                    # in background
+                    self.refresh_task = asyncio.create_task(self._endpoints_health_check(**kwargs))
+                else:
+                    # on startup do this in foreground
+                    await self._endpoints_health_check(**kwargs)
 
     async def _endpoints_health_check(self, **kwargs):
         """Gets the database account for each endpoint.
 
         Validating if the endpoint is healthy else marking it as unavailable.
         """
-        all_endpoints = [self.location_cache.read_regional_endpoints[0]]
-        all_endpoints.extend(self.location_cache.write_regional_endpoints)
-        count = 0
-        for endpoint in all_endpoints:
-            count += 1
-            if count > 3:
+        endpoints_attempted = set()
+        database_account, endpoint = await self._GetDatabaseAccount(**kwargs)
+        endpoints_attempted.add(endpoint)
+        self.location_cache.perform_on_database_account_read(database_account)
+        # should use the endpoints in the order returned from gateway and only the ones specified in preferred locations
+        read_account_regional_routing_contexts_iterator = iter(self.location_cache
+                                                               .account_read_regional_routing_contexts_by_location
+                                                               .values())
+        first_read_regional_routing_context = None
+        # find first read endpoint from gateway and in preferred locations
+        for read_regional_routing_context in read_account_regional_routing_contexts_iterator:
+            if read_regional_routing_context in self.location_cache.read_regional_routing_contexts:
+                first_read_regional_routing_context = read_regional_routing_context
                 break
-            try:
-                await self.client._GetDatabaseAccountCheck(endpoint.get_current(), **kwargs)
-            except (exceptions.CosmosHttpResponseError, AzureError):
-                if endpoint in self.location_cache.read_regional_endpoints:
-                    self.mark_endpoint_unavailable_for_read(endpoint.get_current(), False)
-                if endpoint in self.location_cache.write_regional_endpoints:
-                    self.mark_endpoint_unavailable_for_write(endpoint.get_current(), False)
-                    endpoint.swap()
+        if first_read_regional_routing_context:
+            regional_routing_contexts = [first_read_regional_routing_context]
+        else:
+            regional_routing_contexts = []
+        # add write endpoints from gateway and in preferred locations
+        write_regional_writing_contexts = [endpoint for endpoint in
+                                           self.location_cache
+                                           .account_write_regional_routing_contexts_by_location.values()
+                                           if endpoint in self.location_cache.write_regional_routing_contexts]
+        regional_routing_contexts.extend(write_regional_writing_contexts)
+        success_count = 0
+        for regional_routing_context in regional_routing_contexts:
+            if regional_routing_context.get_primary() not in endpoints_attempted:
+                # health check continues until 2 successes or all endpoints are checked
+                if success_count >= 2:
+                    break
+                endpoints_attempted.add(regional_routing_context.get_primary())
+                try:
+                    await self.client._GetDatabaseAccountCheck(regional_routing_context.get_primary(), **kwargs)
+                    success_count += 1
+                    self.location_cache.mark_endpoint_available(regional_routing_context.get_primary())
+                except (exceptions.CosmosHttpResponseError, AzureError):
+                    if regional_routing_context in self.location_cache.read_regional_routing_contexts:
+                        self.mark_endpoint_unavailable_for_read(regional_routing_context.get_primary(), False)
+                    if regional_routing_context in self.location_cache.write_regional_routing_contexts:
+                        self.mark_endpoint_unavailable_for_write(regional_routing_context.get_primary(), False)
         self.location_cache.update_location_cache()
 
-    async def _GetDatabaseAccount(self, **kwargs):
+    async def _GetDatabaseAccount(self, **kwargs) -> Tuple[DatabaseAccount, str]:
         """Gets the database account.
 
         First tries by using the default endpoint, and if that doesn't work,
         use the endpoints for the preferred locations in the order they are
         specified, to get the database account.
-        :returns: A `DatabaseAccount` instance representing the Cosmos DB Database Account.
-        :rtype: ~azure.cosmos.DatabaseAccount
+        :returns: A `DatabaseAccount` instance representing the Cosmos DB Database Account
+        and the endpoint that was used for the request.
+        :rtype: tuple of (~azure.cosmos.DatabaseAccount, str)
         """
         try:
             database_account = await self._GetDatabaseAccountStub(self.DefaultEndpoint, **kwargs)
             self._database_account_cache = database_account
-            return database_account
+            return database_account, self.DefaultEndpoint
         # If for any reason(non-globaldb related), we are not able to get the database
         # account from the above call to GetDatabaseAccount, we would try to get this
         # information from any of the preferred locations that the user might have
@@ -161,14 +202,17 @@ class _GlobalEndpointManager(object):
         # until we get the database account and return None at the end, if we are not able
         # to get that info from any endpoints
         except (exceptions.CosmosHttpResponseError, AzureError):
+            self.mark_endpoint_unavailable_for_read(self.DefaultEndpoint, False)
+            self.mark_endpoint_unavailable_for_write(self.DefaultEndpoint, False)
             for location_name in self.PreferredLocations:
                 locational_endpoint = LocationCache.GetLocationalEndpoint(self.DefaultEndpoint, location_name)
                 try:
                     database_account = await self._GetDatabaseAccountStub(locational_endpoint, **kwargs)
                     self._database_account_cache = database_account
-                    return database_account
+                    return database_account, locational_endpoint
                 except (exceptions.CosmosHttpResponseError, AzureError):
-                    pass
+                    self.mark_endpoint_unavailable_for_read(locational_endpoint, False)
+                    self.mark_endpoint_unavailable_for_write(locational_endpoint, False)
             raise
 
     async def _GetDatabaseAccountStub(self, endpoint, **kwargs):
@@ -180,3 +224,12 @@ class _GlobalEndpointManager(object):
         :rtype: ~azure.cosmos.DatabaseAccount
         """
         return await self.client.GetDatabaseAccount(endpoint, **kwargs)
+
+    async def close(self):
+        # cleanup any running tasks
+        if self.refresh_task:
+            self.refresh_task.cancel()
+            try:
+                await self.refresh_task
+            except (Exception, CancelledError) : #pylint: disable=broad-exception-caught
+                pass